# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dinas PMD Katingan API Documentation",
        "description": "RESTful API untuk Website Dinas Pemberdayaan Masyarakat dan Desa Kabupaten Katingan.\n *                   API ini menyediakan akses terhadap data berita, agenda, galeri, dokumen, dan struktur organisasi.",
        "contact": {
            "name": "Dinas PMD Katingan",
            "email": "info@pmdkatingan.go.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server (Configurable via .env)"
        }
    ],
    "paths": {
...
        "/api/v1/agenda": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Get list of active agenda",
                "description": "Returns paginated list of active agenda/events with optional date range filter",
                "operationId": "getAgendaList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Filter from date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Filter to date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "upcoming",
                        "in": "query",
                        "description": "Get only upcoming events (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/agenda/{id}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Get agenda detail",
                "description": "Returns detailed information of a specific agenda/event",
                "operationId": "getAgendaDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agenda ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Agenda not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token obtained from login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}